Add LuckyWeb::Exposeable
Closes #51

I don't like that you need to include it. I'm trying to figure out if
there's a better way to do this so you can inherit exposures, but I
can't think of anything quite yet.

I could have made it a module like LuckyWeb::Page, but then the problem
is that in LuckyWeb::Route I can't restrict by a module type e.g.
`LuckyWeb::Action.module`. This is a bit of a bummer. For now I'll move
forward with this and hopefully I can figure out a better solution in
the future.