Add splash page
Closes #75

Very simple at first. Might have a few links with instructions later.

All CSS and images should be available on the page so there is NOTHING
to delete, and nothing to serve with webpack. The page should just work.

Lato is loaded over the web, but it falls back to a sans-serif font that
works fine if you aren't connected.