Set document charset and lang on welcome page
This sets the document charset on the welcome page and prevents the browser warning:
"The character encoding of the HTML document was not declared. The
document will render with garbled text in some browser configurations if
the document contains characters from outside the US-ASCII range. The
character encoding of the page must be declared in the document or in
the transfer protocol."