Show helpful error message when using `p`
There isn't a good reason to use `p` in a web page other than to debug something. So we should add a p macro to `LuckyWeb::Page` that accepts any number of arguments and let's the user know they should use para if they want a paragraph tag and pp if they want to debug. pp is better anyway because it also outputs the code that it's printing to standard out. It's super handy!

Fixes #143