Add checkbox input helper
This is inspired by the Rails and SimpleForm checkbox helpers. There
will always be a hidden field, that transmits the unchecked value.
Otherwise the form would just omit the value, if the checkbox isn't
checked. If the user provides a value for `unchecked_value` we just use
that one. If none given it defaults to "0".

Closes #19