Change routeable's "route" to "with"
When routing to a particular action, an action class is specified. Some
actions also need parameters passed to them. That used to look like:

```crystal

link "edit", to: MyResource::Index

link "edit", to: MyResource::Edit.route(some_params)
```

The `route` method was a confusing here though. You either pass an
action class or the result of it's route method? Too weird. Several
options were discussed:

* `link "edit", to: MyResource::Edit(id: something.id)` - Won't work
syntactically
* `link "edit", to: MyResource::Edit, id: something.id` - confusing
because it could be an html attribute or params
* `link "edit", to: MyResource::Edit.with_params(id: something.id)` -
Too
long given the amount it needs to be typed while using lucky
* `link "edit", to: MyResource::Edit.for(id: something.id)` - Possible.
Some debate was had on if that's the best preposition

In the end, `with` was chosen for its brevity, ease of understanding in
context, and it's ability to convey what needs to be passed to it.