Revert "with" to "route" and pseudo-alias it instead
We decided that in some cases, `with` was a weird choice by itself. If
you need the actual route (an HTTP request method like GET, POST, etc
and a path string) you'd have to call `MyResource::Index.with` which is
too weird.

The compromise is to have both! Use route by itself to return a route
helper that doesn't need params: `MyResource::Index.route` and use
`with` when you need a route helper with params:
`MyResource::Show.with(some_id)`. With just passes its arguments to
route.

The exception to that is if you try to call `with` with no params, it
raises an error. This will hopefully keep the intent of each method
very clean.