Change routeable's "route" to "with" (#179)
* Change routeable's "route" to "with"

When routing to a particular action, an action class is specified. Some
actions also need parameters passed to them. That used to look like:

```crystal

link "edit", to: MyResource::Index

link "edit", to: MyResource::Edit.route(some_params)
```

The `route` method was a confusing here though. You either pass an
action class or the result of it's route method? Too weird. Several
options were discussed:

* `link "edit", to: MyResource::Edit(id: something.id)` - Won't work
syntactically
* `link "edit", to: MyResource::Edit, id: something.id` - confusing
because it could be an html attribute or params
* `link "edit", to: MyResource::Edit.with_params(id: something.id)` -
Too
long given the amount it needs to be typed while using lucky
* `link "edit", to: MyResource::Edit.for(id: something.id)` - Possible.
Some debate was had on if that's the best preposition

In the end, `with` was chosen for its brevity, ease of understanding in
context, and it's ability to convey what needs to be passed to it.

* Revert "with" to "route" and pseudo-alias it instead

We decided that in some cases, `with` was a weird choice by itself. If
you need the actual route (an HTTP request method like GET, POST, etc
and a path string) you'd have to call `MyResource::Index.with` which is
too weird.

The compromise is to have both! Use route by itself to return a route
helper that doesn't need params: `MyResource::Index.route` and use
`with` when you need a route helper with params:
`MyResource::Show.with(some_id)`. With just passes its arguments to
route.

The exception to that is if you try to call `with` with no params, it
raises an error. This will hopefully keep the intent of each method
very clean.