Fix redirect and response status
Close #187

The problem was that you were forced to pass a status to response, so if
you set the status code on the context and then printed a
`LuckyWeb::Response` the original status would be overwritten.

This changes it so that the status can be `nil` on `LuckyWeb::Response`
and it defaults to the one already set on the context, or lastly
defaults to 200.