Make `before` and `after` hooks work with `action`
The problem was that the generated `call` method was not shared, so it
worked for `put/post/get/delete` but not for `action/nested_action`.
This adds a simple test to catch this, and extracts a
`setup_call_method` that is shared between the different ways of setting
up an action.