Introduce LuckyWeb::AllowedInTags for HTML tags
Until now we only did allow Strings for content in HTML tags. Since
people do want to write Int32s and Int64 (and potentially other types as
well) to HTML, we created a new module called `LuckyWeb::AllowedInTags`.

This type is allowed for HTML tags and will be written to HTML with
`to_s`.

Closes #192