Generate Action contents with Teeplate
The Action generator was halfway between using ECR and Teeplate to
render the contents of the Action file.   As such, when generating a new
Action, the file path and name were correct, but file itself just
contained the stringified ActionGenerator object (e.g.
<ActionGenerator:312>).

Switch to using the Teeplate `#render` method on ActionGenerator (now
ActionTemplate) to create a new Action.