Raise friendly error with invalid action
An action must return a response object to be valid. If a developer
fails to do so, the current error message isn't very clear on what needs
to be fix.  Change the error message to explicitly state that one should
return a response from the action, and suggest methods to do so.