More consistent flash handling
Fixes #277

The previous implementation tracked which keys were read/not read. This
ended up being problematic because sometimes keys weren't marked as read
and so they would stay in the flash forever.

Instead, the new implementation only passes the flashes to the next
pages and then removes the rest.

You can use `now` to only set a message for the current request.

All the other hash methods are available because they are forwarded to
`all_messages`.

The `[]` and `[]=` are overridden to work only with symbols and strings.

This also adds a few breaking changes:

* Flash#notice -> success
* Flash#alert -> danger

And a new method

* Flash#info