Hk button helper generates button (#350)
* The 'button' helper now results in a <button/> being rendered.

Previously, the button helper confusingly rendered <a/> links,
likely a result of copy and paste :)

* Rework 'button' helper. Buttons are of only three possible types.

The previous implementation used buttons as links, which is not what
they are intended for. They don't have a href attribute.

* Fix formatting

* Remove button helpers, add to base tags instead.

Buttons were handled the wrong way (like links). However, they
are special, being of type submit, reset or button.