Remove incorrect error message (#459)
Closes #437

It seems this may be an issue in Crystal, but I need more time to track
it down. For now, I'm going to remove this special error-handling and
add it back later. That's better than giving people incorrect
instructions. Note that a compile-time error will still be raised, it
will just be a generic Crystal error.

Follow-up issue: https://github.com/luckyframework/lucky/issues/460