Add support for HTTP responses with files
Similar to Rails' ActionController [send_file] method, this commit adds
support for the renderable class to not only render text, but to also
respond with the contents of a file.

This also allows for designating whether a file should be downloaded
upon hitting the action, or rendered inline (for images, or maybe a text
file).

[send_file]: https://edgeapi.rubyonrails.org/classes/ActionController/DataStreaming.html#method-i-send_file