CookieJar is now an extension of HTTP::Cookies
This is so other attributes on cookies can be set. For example if you
want to set an expiration for a cookie, you need access to the raw
cookie, or Lucky would have to essentially reimplement cookies work and
they work fine for the most part.

The only thing we'll be doing is encrypting/decrypting all their
values. That's up next.

I also removed the key setting because there's not a single key for
cookies. There will be a single key for the session though.