Change interface to the CookieJar
This was changed primarily so that we could abstract away the actual
cookies object in case it changed, and also so you can only set/get
cookies with a symbol or string. HTTP::Cookies lets you use any object
as a key which we didn't feel was safe enough.

abstract away the actual cookies object in case it changed, and also so
you can only set/get cookies with a symbol or string. HTTP::Cookies
lets you use any object as a key which we didn't feel was safe enough.
'el-cookie-refactor' on '92d45bb'.