Set the retrieve the session from cookies
The session is now a single cookie with a name ending in "_session" and
starting with the name of the app (eventually). It is encoded with the
cookies when they are written to the response and can be fetched from
the cookies in the request with their preset key.

Changing the key will invalidate the session because it will start
fresh.