Introduce NullCookie
NullCookie is a HTTP::Cookie-like object that is meant to have a
similar API and respond to most of the same getters. For each value, it
returns nil or false.

It's now used when retrieving a cookie with the get?() method. If there
is no cookie with the specified key, it will return this NullCookie.
That way people can call all the normal methods on the cookie without
the app crashing.