Create FlashHandler to export the flash to the session
This is similar to the session handler, but uses the session as the
store for the Flash instead of a cookie. The session of course
eventually turns into a cookie, but the Flash handler doesn't care.

The flash handler also only submits the "next" flash to the json, which
gets read in as the "now" flash. This ensures that the flash only
sticks around for one request/response cycle.