Set far-future expiration dates on cookies
I chose 1 year which seems like a reasonable default for most apps.
It's not yet configurable unless you manually grab the cookie and set
it yourself.

I also had to change the encryptor around a little bit so it used the
expiration date that I set. This also caused me to change a test in the
session handler spec. Now when you call the SessionHandler, it will
encrypt all the cookies on the context as opposed to writing new ones
to a response. Because of this, the cookies I was reading to test if
they could be transferred from session to session weren't readable
because they got encrypted.

But, it turns out that I don't want to call the SessionHandler with the
new context anyway since in a real world scenario the cookies will be
accessed before the session handler really finished.