Dirty hack to fix cookie expiration spec
This works when I test just this file but not in the full suite. My
guess is because the time set in the cookie jar is a constant that is
set at compile time, which ends taking more than 1 second to compile. 1
minute difference for a year seems reasonable for now.

I tried to get timecop.cr to work but even that showed differences in
time of more than one second.