Don't write cookies or session if they haven't changed
This tracks a @changed instance variable on the CookieJar and
SessionCookie. It starts as false and changes to true any time the
`set` method is called.

The reason we want to check if any cookies have changes is sending the
Set-Cookie headers back in a response makes the response larger. If no
cookies have changed, we can save bandwidth and gain speed by just not
sending them in the request.

If the session hasn't been saved, it doesn't get set on the cookie. If
they session has changed however, it will set a new cookie which marks
the cookies as having changed and therefore will send the Set-Cookie
header in the response.