Raise error if cookie size is over 4096 bytes
Most browsers don't support cookie headers that carry more than 4096
bytes of data. This will raise an error if someone is trying to write
cookies that are bigger than that.

The tricky thing is the encryptor will give a slightly different output
every time it's written, so you could theoretically have some cookies
that are 4090 bytes one write, and 4097 bytes the next which will trip
this error. Not quite sure what to do about that.