Use span for cookie expiration instead of time
The problem with using a Time as a constant is that the date is based
on the running crystal process. So for example if you boot up the app
on Jan 1, 2018 1 year from now will be Jan 1 2019. If the app continues
to run, that 1 year from now date will not change.

So instead I'm storing it as a Time::Span (or Time::MonthSpan) and
converting it to a time just as I set the cookie so it will calculate
from the time the cookie is set, not the time the app was booted. In
addition, I made it a habitat setting so it could be configured by the
developer.

I could have used a method too, but that wouldn't have been
configurable with habitat.