Require session cookie to be set with a Cookie
It was a bit risky to have the session be a NullCookie by default, so
instead we require it to use an HTTP cookie.

However, now the cookie needed a way to bootstrap itself. I added a
session_cookie method to CookeJar that would return the stored cookie,
or make a new blank one ready for the session to eat (I mean initialize
with).