Fix `BaseHTTPClient` params
The underlying `HTTP::Client` was expecting a host, but we were
providing it a base URI (which is the combined host and port).

It was also deriving its default value from the `RouteHelper`, but the
requests are expected to be made to the `Server` *first*.

This also enables the host to be configured without overriding the
client.