distance_of_time_in_words returns a string instead of text element (#781)
* distance_of_time_in_words returns a string instead of text element

* Documentation for simple_format

* pluralize, word_wrap, to_sentence, and cycle return strings

The goal is to make these easier to use inside other tags. The problem
was that these would all output to the IO, so if you tried to use these
in an element, it would print out the whole page HTML.

* Return a string and add docs for highlight text helper

* Allow excerpt to be used inside tags and add docs

* Make truncate_text public and add truncate docs

* Remove extraneous string conversions in text helper tests

* Update changelog

* View helper methods that don't return String should return Nil

This change also made the tests have to change since we can no longer
grab the output of the text helper to a string and then compare it with
another string.