More reliable and configurable asset helpers
Closes #717

It also only looked for a manifest but not a particular file. The
problem with that is that you may have an old manifest that is in the
process of being updated. That means Lucky incorrectly thinks the asset
doesn't exist, when in reality the asset compilation process just hasn't
finished and therefore hasn't updated the manifest.

Instead we look for particular assets. So even if the manifest exists it
will retry until the new asset is there or it times out

---

Note the ENV vars must be available at compile time. This means setting
them in a .env file will not work. Instead add them to your command in
Procfile.dev.

    web: ASSET_HELPER_TIMEOUT_IN_SECONDS=15 lucky watch