Add spec showing that HttpMethodOverrideHandler fails on invalid json (#854)
Because part of this handler looks at the request body, it can fail if the json is malformed.  I believe this is totally acceptable behavior.  The reason I point this out is because in lucky-cli, this handler is put before the ErrorHandler ([reference](https://github.com/luckyframework/lucky_cli/blob/master/src/web_app_skeleton/src/app_server.cr.ecr\#L5-L15)) when it probably shouldn't be. This is most likely the cause of https://github.com/luckyframework/lucky/issues/852