Fix issues with accepted_formats macro
The _accepted_formats class property was not inherited by child actions
so when the verified formats method was called, it thought it accepted
every format. That also meant that 'clients_desired_format' was not
called.  So anything went through and the format was never cached. So
'Errors::Show' did not know what format to use.

Instead we generate a method instead of using a class property. This
ensure the method is properly inherited.

We also explicitly cache the desired format at the start of the action