Add compiling spinner to watch task
This does not work yet. We need to point the other shards in Lucky to
use LuckyCli master. Since they don't I just brought in the Spinner for
testing.

The spinner works fine when running `lucky watch` directly. But when run
inside of overmind it shows nothing...not sure why